requirements.txt
alabaster==1.0.0
babel==2.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
docutils==0.21.2
feedgen==1.0.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
myst-parser==4.0.1
packaging==25.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-frontmatter==1.1.0
PyYAML==6.0.2
requests==2.32.3
roman-numerals-py==3.1.0
six==1.17.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
urllib3==2.4.0
